feat(build): add additional binaries (#22)
# 📥 Pull Request

## ❓ What are you trying to address

This pull request includes changes to add support for additional
binaries in the build artifacts. The most important changes involve
updates to the build configuration to include FreeBSD and 386
architectures.

## ✨ Description of new changes

Build configuration updates:

*
[`.goreleaser.yml`](diffhunk://#diff-42e26dc67aed8aa3edb2472b4403288c1699fb6dc47419b9a475f0f224fe4689R26-R37):
Updated `goos` to include FreeBSD and `goarch` to include 386. Also
modified the `ignore` section to exclude 386 architecture for Darwin.