docs: update faq in the troubleshooting guide (#28)
# 📥 Pull Request

## ❓ What are you trying to address

This pull request includes updates to the documentation files to improve
clarity and accuracy. The most important changes are corrections to the
phrasing and formatting of instructions related to Fabric Capacity and
workspace management.

## ✨ Description of new changes

Documentation improvements:

*
[`docs/guides/troubleshooting.md`](diffhunk://#diff-05fa5078290d9319b91e520b8d624cd018e97d963be0d0e1cd22ca7e37e899e9L38-R52):
Corrected phrasing to ensure clarity in instructions about assigning
Fabric Capacity and checking user permissions.
*
[`templates/guides/troubleshooting.md.tmpl`](diffhunk://#diff-05fa5078290d9319b91e520b8d624cd018e97d963be0d0e1cd22ca7e37e899e9L38-R52):
Updated template to match changes in the main troubleshooting guide,
ensuring consistent phrasing and formatting.