ci(test): add manual trigger option to test workflow (#38)
# 📥 Pull Request

## ❓ What are you trying to address

This pull request includes updates to the GitHub Actions workflow and
the `README.md` file to enhance project documentation and workflow
capabilities.

## ✨ Description of new changes


Documentation updates:

*
[`README.md`](diffhunk://#diff-b335630551682c19a781afebcf4d07bf978fb1f8ac04c6bf87428ed5106870f5R3):
Added a badge for the test workflow to provide a quick visual indication
of the test status.

Workflow enhancements:

*
[`.github/workflows/test.yml`](diffhunk://#diff-faff1af3d8ff408964a57b2e475f69a6b7c7b71c9978cccc8f471798caac2c88R17):
Added the `workflow_dispatch` trigger to allow manual triggering of the
workflow.