refactor(fabricitem): use generics for properties (#140)
# 📥 Pull Request

## ❓ What are you trying to address

This pull request includes several changes to the
`fabric_spark_job_definition` by adding the extended properties.

Additionally, PR contains internal refactoring for Fabric Items to
handle extended properties using Go Generics. This will be the base for
refactoring of exiting items to use it and for any future items with
properties.

---------

Co-authored-by: Pablo Zaidenvoren <pablozaiden@users.noreply.github.com>