build(deps): update go to 1.23.6 and terraform to 1.10.5 (#236)
# 📥 Pull Request

## ❓ What are you trying to address

This pull request includes updates to the versions of Go and Terraform
across various configuration and documentation files to ensure
consistency and compatibility.

Version updates:

*
[`.devcontainer/devcontainer.json`](diffhunk://#diff-24ad71c8613ddcf6fd23818cb3bb477a1fb6d83af4550b0bad43099813088686L18-R18):
Updated the Go feature version from `1.23.5` to `1.23.6`.
*
[`.devcontainer/devcontainer.json`](diffhunk://#diff-24ad71c8613ddcf6fd23818cb3bb477a1fb6d83af4550b0bad43099813088686L39-R39):
Updated the Terraform feature version from `1.10.4` to `1.10.5`.
*
[`DEVELOPER.md`](diffhunk://#diff-35ea617bfbf0780bfbc554348314c5b11ba478a4d98fd92d33d45cdd54bf326aL111-R116):
Updated the required versions of Go and Terraform to `1.23.6` and
`1.10.5` respectively, and adjusted the version manager instructions
accordingly.
*
[`go.mod`](diffhunk://#diff-33ef32bf6c23acb95f5902d7097b7a1d5128ca061167ec0716715b0b9eeaa5f6L3-R3):
Updated the Go version from `1.23.5` to `1.23.6`.