fix(rs-spark_custom_pool): inconsistent result for dynamic_executor_allocation (#240)
# 📥 Pull Request

fix #239

## ❓ What are you trying to address

This pull request includes changes to address the inconsistent results
for dynamic executor allocation when the `enabled` attribute is set to
false and updates the documentation and schema accordingly.

## ✨ Description of new changes

### Code updates:
* `internal/services/spark/resource_spark_custom_pool.go`: 
* Added new imports for `attr`, `types`, and `superint32validator`
packages.
* Updated the schema for `min_executors` and `max_executors` attributes
to be `Computed` and `Optional`, and added validators to handle the
`dynamic_executor_allocation` `enabled` attribute.

### Documentation updates:
*
[`docs/resources/spark_custom_pool.md`](diffhunk://#diff-e2c5956959615a4e86cc330845da0052ce2316efdf948d7bea320eb2365fbab3R81-R83):
Added an `Optional` section to the documentation for `max_executors` and
`min_executors` attributes.