build(deps): update go, tf and devcontainer (#285)
# 📥 Pull Request

## ❓ What are you trying to address

This pull request includes updates to the development environment
configurations and documentation to ensure compatibility with the latest
versions of Go and Terraform.

Updates to development environment:

*
[`.devcontainer/devcontainer.json`](diffhunk://#diff-24ad71c8613ddcf6fd23818cb3bb477a1fb6d83af4550b0bad43099813088686L12-R18):
Updated the `fish` feature to version 2, the `go` feature to version
1.24.1, and the `terraform` feature to version 1.11.1.
[[1]](diffhunk://#diff-24ad71c8613ddcf6fd23818cb3bb477a1fb6d83af4550b0bad43099813088686L12-R18)
[[2]](diffhunk://#diff-24ad71c8613ddcf6fd23818cb3bb477a1fb6d83af4550b0bad43099813088686L39-R39)

Documentation updates:

*
[`DEVELOPER.md`](diffhunk://#diff-35ea617bfbf0780bfbc554348314c5b11ba478a4d98fd92d33d45cdd54bf326aL111-R116):
Updated the required versions of Go to 1.24.1 and Terraform to 1.11.1 in
the requirements section.

Codebase updates:

*
[`go.mod`](diffhunk://#diff-33ef32bf6c23acb95f5902d7097b7a1d5128ca061167ec0716715b0b9eeaa5f6L3-R3):
Updated the Go version to 1.24.1.