feat: enhance release preparation (#296)
# 📥 Pull Request

## ❓ What are you trying to address

This pull request includes several changes to configuration files and
task definitions to improve the release process and streamline tool
installations. The most important changes include updating the release
preparation process, modifying auto versioning rules, and adding new
tasks for installing the GitHub CLI.

### Release Process Improvements:

*
[`Taskfile.yml`](diffhunk://#diff-cd2d359855d0301ce190f1ec3b4c572ea690c83747f6df61c9340720e3d2425eL362-L390):
Enhanced the `release:prep` task to include preconditions for ensuring
the user is on the main branch, has a clean working tree, and provides a
valid semver format version. Updated the task to use the `VERSION`
variable instead of `SEMVER` and included additional commands for
creating and merging pull requests.

### Versioning Rules:

*
[`.changie.yaml`](diffhunk://#diff-9f1c8d07b49e66598bb525c145a5d7524fef447d6993b72ccb378e70ebf922f3L14-R32):
Changed the auto versioning rules for different labels to use `minor`
and `patch` instead of commenting them out. This ensures that version
increments happen automatically based on the type of change.

### Tool Installations:

*
[`Taskfile.yml`](diffhunk://#diff-cd2d359855d0301ce190f1ec3b4c572ea690c83747f6df61c9340720e3d2425eR553-R570):
Added a new task `install:gh` for installing the GitHub CLI on different
platforms (Windows, macOS, and Linux). This ensures that the GitHub CLI
is available for use in other tasks.

### Cleanup and Removal:

*
[`.semver.yaml`](diffhunk://#diff-e41972a4655bc6309bdc509296f4111a35952b8d3b554f7752396f05bc95f57cL1-L4):
Removed the file as it is no longer needed with the updated release
process.
*
[`Taskfile.yml`](diffhunk://#diff-cd2d359855d0301ce190f1ec3b4c572ea690c83747f6df61c9340720e3d2425eL131):
Removed the `install:semver` task and references to `semver` in other
tasks, as the semver tool is no longer used.
[[1]](diffhunk://#diff-cd2d359855d0301ce190f1ec3b4c572ea690c83747f6df61c9340720e3d2425eL131)
[[2]](diffhunk://#diff-cd2d359855d0301ce190f1ec3b4c572ea690c83747f6df61c9340720e3d2425eR144)