feat(task): add GitHub CLI authentication check to Taskfile (#304)
# 📥 Pull Request

## ❓ What are you trying to address

This pull request includes a small change to the `Taskfile.yml` file.
The change adds a new shell command to check the GitHub CLI
authentication status and provides a message if the user is not
authenticated.

*
[`Taskfile.yml`](diffhunk://#diff-cd2d359855d0301ce190f1ec3b4c572ea690c83747f6df61c9340720e3d2425eR387-R389):
Added a shell command to check GitHub CLI authentication status with a
corresponding message for unauthenticated users.