refactor(logging): move to client package package (#314)
# 📥 Pull Request

## ❓ What are you trying to address

This pull request introduces enhancements to the logging functionality
for the Fabric SDK, including the addition of advanced logging options
and corresponding documentation updates. The most important changes
include the introduction of new environment variables for logging
configuration, the implementation of functions to handle these
configurations, and the addition of comprehensive tests to ensure the
correctness of these new features.

These changes enhance the ability to debug issues by providing more
granular control over logging behavior, while ensuring that sensitive
information is handled with care.