chore(repo): update issue templates (#323)
This pull request includes several updates to the GitHub issue templates
to improve clarity and organization. The most important changes include
adding a `type` field to various templates, updating descriptions for
better clarity, and reorganizing the contact links.

### Updates to Issue Templates:

*
[`.github/ISSUE_TEMPLATE/bug_report.yml`](diffhunk://#diff-637f7b97bba458badb691a1557c3d4648686292e948dbe3e8360564378b653efR8):
Added a `type` field with the value "Bug" to the bug report template.
*
[`.github/ISSUE_TEMPLATE/feature_request.yml`](diffhunk://#diff-c6b098646bf32c644234bec14c2675ea2a3d9c2dc2df450a25aa0e7ff02323cdL5-R8):
Added a `type` field with the value "Feature" and updated the
description for better clarity.
*
[`.github/ISSUE_TEMPLATE/tfprovider_data_source_request.yml`](diffhunk://#diff-3363b4ebd01ea6ad113865c5b2d033b770df2b39203cc273faaa6e2d463afa16L5-R8):
Added a `type` field with the value "Feature" and updated the
description for better clarity.
*
[`.github/ISSUE_TEMPLATE/tfprovider_docs_request.yml`](diffhunk://#diff-b653c577aa3db5c19639da4d289d6b7f3fc32efabad5794c7c6c624a36b7243aL5-R8):
Added a `type` field with the value "Feature" and updated the
description for better clarity.
*
[`.github/ISSUE_TEMPLATE/tfprovider_resource_request.yml`](diffhunk://#diff-c4eaf517a5af4e0890a9e5dc89340340a6d2f24988f55e068fdf8ae21b105e77L5-R8):
Added a `type` field with the value "Feature" and updated the
description for better clarity.

### Reorganization of Contact Links:

*
[`.github/ISSUE_TEMPLATE/config.yml`](diffhunk://#diff-1c0d972ee49103af56fd608a77a28e1eb12f6908f263f0f183d46868fdcd8ea2L6-R18):
Reorganized the contact links to include new categories such as
"Question" and "Community Support and Discussions" with updated URLs and
descriptions.

### Removal of Question Template:

*
[`.github/ISSUE_TEMPLATE/question.yml`](diffhunk://#diff-b6087b0ddc2695a819e8199639ea5be88e1352a9a10774c7183eefa682e2dc76L1-L46):
Removed the question template entirely.